worker-service/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.33.0